+++ Data/Rettelser.txt
@@ -74,7 +74,4 @@ erstat;5710441340829;5710441350941
 erstat;5710441340850;5710441350897
 erstat;5710441340874;5710441350910
 erstat;5710441340898;5710441350934
-# erstat;5710441340836;5710441350958
-flere;5710441340836;5710441350958;5710441350934;5710441350910
-flere;5710441350958;5710441350934;5710441350910
-
+erstat;5710441340836;5710441350958
